refactor: eliminate controller and path related duplications
