+++ lib/assertion-visitor.js
@@ -10,6 +10,7 @@ const syntax = estraverse.Syntax;
 const EspowerLocationDetector = require('espower-location-detector');
 const toBeSkipped = require('./rules/to-be-skipped');
 const toBeCaptured = require('./rules/to-be-captured');
+const { getParentNode } = require('./controller-utils');
 const canonicalCodeOptions = {
   format: {
     indent: {
@@ -256,11 +257,6 @@ const findEnclosingFunction = (parents) => {
     }
   }
   return null;
-};
-
-const getParentNode = (controller) => {
-  const parents = controller.parents();
-  return parents[parents.length - 1];
 };
 
 module.exports = AssertionVisitor;
