+++ lib/instrumentor.js
@@ -6,6 +6,7 @@ const escope = require('escope');
 const createPatternMatchers = require('../lib/create-pattern-matchers');
 const AssertionVisitor = require('./assertion-visitor');
 const Transformation = require('./transformation');
+const { getCurrentKey } = require('./controller-utils');
 const isSpreadElement = (node) => node.type === 'SpreadElement';
 const withoutMatcherAndIndex = (p) => {
   const res = Object.assign({}, p);
@@ -51,8 +52,7 @@ class Instrumentor {
           scopeStack.push(scopeManager.acquire(currentNode));
         }
         const controller = this;
-        const path = controller.path();
-        const currentKey = path ? path[path.length - 1] : null;
+        const currentKey = getCurrentKey(controller);
         if (assertionVisitor) {
           if (assertionVisitor.toBeSkipped(controller)) {
             skipping = true;
