+++ lib/rules/to-be-captured.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { Syntax } = require('estraverse');
+const { getParentNode, getCurrentKey } = require('../controller-utils');
 const caputuringTargetTypes = [
     // Syntax.Property,
     Syntax.ObjectExpression,
@@ -36,16 +37,6 @@ const isChildOfTaggedTemplateExpression = (parentNode) => {
 const isYieldOrAwaitArgument = (parentNode, currentKey) => {
     // capture the yielded/await result, not the promise
     return (parentNode.type === Syntax.YieldExpression || parentNode.type === Syntax.AwaitExpression) && currentKey === 'argument';
-};
-
-const getParentNode = (controller) => {
-    const parents = controller.parents();
-    return parents[parents.length - 1];
-};
-
-const getCurrentKey = (controller) => {
-    const path = controller.path();
-    return path ? path[path.length - 1] : null;
 };
 
 const toBeCaptured = (controller) => {
