+++ lib/rules/to-be-skipped.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { Syntax } = require('estraverse');
+const { getParentNode, getCurrentKey } = require('../controller-utils');
 const supportedNodeTypes = require('./supported-node-types');
 
 const isLeftHandSideOfAssignment = (parentNode, currentKey) => {
@@ -47,16 +48,6 @@ const isTypeOfOrDeleteUnaryExpression = (currentNode, parentNode, currentKey) =>
 
 const isSupportedNodeType  = (node) => {
     return supportedNodeTypes.indexOf(node.type) !== -1;
-};
-
-const getParentNode = (controller) => {
-    const parents = controller.parents();
-    return parents[parents.length - 1];
-};
-
-const getCurrentKey = (controller) => {
-    const path = controller.path();
-    return path ? path[path.length - 1] : null;
 };
 
 const toBeSkipped = (controller) => {
